projects/prediction_server/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
-e git+https://github.com/JNeuvonen/backtest-platform.git@bd19ecd960db1b01ed98cf32b0670ea612d600c9#egg=common_python&subdirectory=projects/common_python
dateparser==1.2.0
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.3
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pluggy==1.4.0
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pytest==8.1.1
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
websockets==12.0
yarl==1.9.4
